+++ api/discovery/resource.go
@@ -1,0 +1,1 @@
+package discovery
