+++ service/discovery/azure/azure_clients.go
