+++ service/discovery/azure/backup_discover.go
@@ -157,69 +157,3 @@ func (d *azureDiscovery) discoverBackupInstances(resourceGroup, vaultName string
 
 	return list.Value, nil
 }
-
-func (d *azureDiscovery) handleInstances(vault *armdataprotection.BackupVaultResource, instance *armdataprotection.BackupInstanceResource) (resource voc.IsCloudResource, err error) {
-	if vault == nil || instance == nil {
-		return nil, ErrVaultInstanceIsEmpty
-	}
-
-	raw, err := voc.ToStringInterface([]interface{}{instance, vault})
-	if err != nil {
-		log.Error(err)
-	}
-
-	if *instance.Properties.DataSourceInfo.DatasourceType == "Microsoft.Storage/storageAccounts/blobServices" {
-		resource = &voc.ObjectStorage{
-			Storage: &voc.Storage{
-				Resource: &voc.Resource{
-					ID:           voc.ResourceID(*instance.ID),
-					Name:         *instance.Name,
-					CreationTime: 0,
-					GeoLocation: voc.GeoLocation{
-						Region: *vault.Location,
-					},
-					Labels:    nil,
-					ServiceID: d.csID,
-					Type:      voc.ObjectStorageType,
-					Parent:    resourceGroupID(instance.ID),
-					Raw:       raw,
-				},
-			},
-		}
-	} else if *instance.Properties.DataSourceInfo.DatasourceType == "Microsoft.Compute/disks" {
-		resource = &voc.BlockStorage{
-			Storage: &voc.Storage{
-				Resource: &voc.Resource{
-					ID:           voc.ResourceID(*instance.ID),
-					Name:         *instance.Name,
-					ServiceID:    d.csID,
-					CreationTime: 0,
-					Type:         voc.BlockStorageType,
-					GeoLocation: voc.GeoLocation{
-						Region: *vault.Location,
-					},
-					Labels: nil,
-					Parent: resourceGroupID(instance.ID),
-					Raw:    raw,
-				},
-			},
-		}
-	}
-
-	return
-}
-
-// backupsEmptyCheck checks if the backups list is empty and returns voc.Backup with enabled = false.
-func backupsEmptyCheck(backups []*voc.Backup) []*voc.Backup {
-	if len(backups) == 0 {
-		return []*voc.Backup{
-			{
-				Enabled:         false,
-				RetentionPeriod: -1,
-				Interval:        -1,
-			},
-		}
-	}
-
-	return backups
-}
