--- service/discovery/azure/storage.go
@@ -1,770 +1,0 @@
-// Copyright 2021 Fraunhofer AISEC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-//           $$\                           $$\ $$\   $$\
-//           $$ |                          $$ |\__|  $$ |
-//  $$$$$$$\ $$ | $$$$$$\  $$\   $$\  $$$$$$$ |$$\ $$$$$$\    $$$$$$\   $$$$$$\
-// $$  _____|$$ |$$  __$$\ $$ |  $$ |$$  __$$ |$$ |\_$$  _|  $$  __$$\ $$  __$$\
-// $$ /      $$ |$$ /  $$ |$$ |  $$ |$$ /  $$ |$$ |  $$ |    $$ /  $$ |$$ | \__|
-// $$ |      $$ |$$ |  $$ |$$ |  $$ |$$ |  $$ |$$ |  $$ |$$\ $$ |  $$ |$$ |
-// \$$$$$$\  $$ |\$$$$$   |\$$$$$   |\$$$$$$  |$$ |  \$$$   |\$$$$$   |$$ |
-//  \_______|\__| \______/  \______/  \_______|\__|   \____/  \______/ \__|
-//
-// This file is part of Clouditor Community Edition.
-
-package azure
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"strings"
-
-	"clouditor.io/clouditor/api/discovery"
-	"clouditor.io/clouditor/internal/constants"
-	"clouditor.io/clouditor/internal/util"
-	"clouditor.io/clouditor/voc"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/cosmos/armcosmos"
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/sql/armsql"
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/storage/armstorage"
-)
-
-var (
-	ErrEmptyStorageAccount        = errors.New("storage account is empty")
-	ErrMissingDiskEncryptionSetID = errors.New("no disk encryption set ID was specified")
-	ErrBackupStorageNotAvailable  = errors.New("backup storages not available")
-)
-
-// discoverCosmosDB discovers Cosmos DB accounts
-func (d *azureDiscovery) discoverCosmosDB() ([]voc.IsCloudResource, error) {
-	var (
-		list []voc.IsCloudResource
-		err  error
-	)
-
-	// initialize Cosmos DB client
-	if err := d.initCosmosDBClient(); err != nil {
-		return nil, err
-	}
-
-	// Discover Cosmos DB
-	err = listPager(d,
-		d.clients.cosmosDBClient.NewListPager,
-		d.clients.cosmosDBClient.NewListByResourceGroupPager,
-		func(res armcosmos.DatabaseAccountsClientListResponse) []*armcosmos.DatabaseAccountGetResults {
-			return res.Value
-		},
-		func(res armcosmos.DatabaseAccountsClientListByResourceGroupResponse) []*armcosmos.DatabaseAccountGetResults {
-			return res.Value
-		},
-		func(dbAccount *armcosmos.DatabaseAccountGetResults) error {
-			cosmos, err := d.handleCosmosDB(dbAccount)
-			if err != nil {
-				return fmt.Errorf("could not cosmos db accounts: %w", err)
-			}
-			log.Infof("Adding Cosmos DB account '%s", *dbAccount.Name)
-			list = append(list, cosmos...)
-
-			return nil
-		})
-	if err != nil {
-		return nil, err
-	}
-
-	return list, nil
-}
-
-func (d *azureDiscovery) handleCosmosDB(account *armcosmos.DatabaseAccountGetResults) ([]voc.IsCloudResource, error) {
-	var (
-		atRestEnc voc.IsAtRestEncryption
-		err       error
-		list      []voc.IsCloudResource
-	)
-
-	// initialize Cosmos DB client
-	if err = d.initCosmosDBClient(); err != nil {
-		return nil, err
-	}
-
-	// Check if KeyVaultURI is set for Cosmos DB account
-	// By default the Cosmos DB account is encrypted by Azure managed keys. Optionally, it is possible to add a second encryption layer with customer key encryption. (see https://learn.microsoft.com/en-us/azure/cosmos-db/how-to-setup-customer-managed-keys?tabs=azure-portal)
-	if account.Properties.KeyVaultKeyURI != nil {
-		atRestEnc = &voc.CustomerKeyEncryption{
-			AtRestEncryption: &voc.AtRestEncryption{
-				Enabled: true,
-				// Algorithm: algorithm, //TODO(anatheka): How do we get the algorithm? Are we available to do it by the related resources?
-			},
-			KeyUrl: util.Deref(account.Properties.KeyVaultKeyURI),
-		}
-	} else {
-		atRestEnc = &voc.ManagedKeyEncryption{
-			AtRestEncryption: &voc.AtRestEncryption{
-				Enabled:   true,
-				Algorithm: AES256,
-			},
-		}
-	}
-
-	// Create Cosmos DB database service voc object for the database account
-	dbService := &voc.DatabaseService{
-		StorageService: &voc.StorageService{
-			NetworkService: &voc.NetworkService{
-				Networking: &voc.Networking{
-					Resource: discovery.NewResource(d,
-						voc.ResourceID(*account.ID),
-						*account.Name,
-						account.SystemData.CreatedAt,
-						voc.GeoLocation{
-							Region: *account.Location,
-						},
-						labels(account.Tags),
-						resourceGroupID(account.ID),
-						voc.DatabaseServiceType,
-						account,
-					),
-				},
-			},
-		},
-	}
-
-	// Add Mongo DB database service
-	list = append(list, dbService)
-
-	// Check account kind and add Mongo DB databases storages
-	switch util.Deref(account.Kind) {
-	case armcosmos.DatabaseAccountKindMongoDB:
-		// Get Mongo databases
-		list = append(list, d.getMongoDBs(account, atRestEnc)...)
-	case armcosmos.DatabaseAccountKindGlobalDocumentDB:
-		log.Infof("%s not yet implemented", armcosmos.DatabaseAccountKindGlobalDocumentDB)
-	case armcosmos.DatabaseAccountKindParse:
-		log.Infof("%s not yet implemented", armcosmos.DatabaseAccountKindParse)
-	default:
-		log.Warningf("Account kind '%s' not yet implemented", util.Deref(account.Kind))
-	}
-
-	return list, nil
-}
-
-// getMongoDBs returns a list of Mongo DB databases for a specific Mongo DB account
-func (d *azureDiscovery) getMongoDBs(account *armcosmos.DatabaseAccountGetResults, atRestEnc voc.IsAtRestEncryption) []voc.IsCloudResource {
-	var (
-		list []voc.IsCloudResource
-		err  error
-	)
-
-	// initialize Mongo DB resources client
-	if err = d.initMongoDResourcesBClient(); err != nil {
-		log.Errorf("error initializing Mongo DB resource client: %v", err)
-		return list
-	}
-
-	// Discover Mongo DB databases
-	serverlistPager := d.clients.mongoDBResourcesClient.NewListMongoDBDatabasesPager(resourceGroupName(util.Deref(account.ID)), *account.Name, &armcosmos.MongoDBResourcesClientListMongoDBDatabasesOptions{})
-	for serverlistPager.More() {
-		pageResponse, err := serverlistPager.NextPage(context.TODO())
-		if err != nil {
-			log.Errorf("%s: %v", ErrGettingNextPage, err)
-			return list
-		}
-
-		for _, value := range pageResponse.Value {
-			// Create Cosmos DB database storage voc object
-			mongoDB := &voc.DatabaseStorage{
-				Storage: &voc.Storage{
-					Resource: discovery.NewResource(d,
-						voc.ResourceID(*value.ID),
-						util.Deref(value.Name),
-						nil, // creation time of database not available
-						voc.GeoLocation{
-							Region: *value.Location,
-						},
-						labels(value.Tags),
-						voc.ResourceID(*account.ID),
-						voc.DatabaseStorageType,
-						account,
-						value),
-
-					AtRestEncryption: atRestEnc,
-				},
-			}
-			list = append(list, mongoDB)
-		}
-	}
-
-	return list
-}
-
-// discoverSqlServers discovers the sql server and databases
-func (d *azureDiscovery) discoverSqlServers() ([]voc.IsCloudResource, error) {
-	var (
-		list []voc.IsCloudResource
-		err  error
-	)
-
-	// initialize SQL server client
-	if err := d.initSQLServersClient(); err != nil {
-		return nil, err
-	}
-
-	// Discover sql server
-	err = listPager(d,
-		d.clients.sqlServersClient.NewListPager,
-		d.clients.sqlServersClient.NewListByResourceGroupPager,
-		func(res armsql.ServersClientListResponse) []*armsql.Server {
-			return res.Value
-		},
-		func(res armsql.ServersClientListByResourceGroupResponse) []*armsql.Server {
-			return res.Value
-		},
-		func(server *armsql.Server) error {
-			db, err := d.handleSqlServer(server)
-			if err != nil {
-				return fmt.Errorf("could not handle sql database: %w", err)
-			}
-			log.Infof("Adding sql database '%s", *server.Name)
-			list = append(list, db...)
-
-			return nil
-		})
-	if err != nil {
-		return nil, err
-	}
-
-	return list, nil
-}
-
-func (d *azureDiscovery) handleSqlServer(server *armsql.Server) ([]voc.IsCloudResource, error) {
-	var (
-		dbList               []voc.IsCloudResource
-		anomalyDetectionList []voc.IsAnomalyDetection
-		dbService            voc.IsCloudResource
-		list                 []voc.IsCloudResource
-	)
-
-	// Get SQL database storages and the corresponding anomaly detection property
-	dbList, anomalyDetectionList = d.getSqlDBs(server)
-
-	// Create SQL database service voc object for SQL server
-	dbService = &voc.DatabaseService{
-		StorageService: &voc.StorageService{
-			NetworkService: &voc.NetworkService{
-				Networking: &voc.Networking{
-					Resource: discovery.NewResource(d,
-						voc.ResourceID(*server.ID),
-						*server.Name,
-						nil, // creation time not available
-						voc.GeoLocation{
-							Region: *server.Location,
-						},
-						labels(server.Tags),
-						resourceGroupID(server.ID),
-						voc.DatabaseServiceType,
-						server,
-					),
-				},
-				// TODO(all): HttpEndpoint
-				TransportEncryption: &voc.TransportEncryption{
-					Enabled:    true,
-					Enforced:   true,
-					TlsVersion: checkTlsVersion(*server.Properties.MinimalTLSVersion),
-				},
-			},
-		},
-		AnomalyDetection: anomalyDetectionList,
-	}
-
-	// Add SQL database service
-	list = append(list, dbService)
-
-	// Add SQL database storages
-	list = append(list, dbList...)
-
-	return list, nil
-}
-
-// getSqlDBs returns a list of SQL databases for a specific SQL account
-func (d *azureDiscovery) getSqlDBs(server *armsql.Server) ([]voc.IsCloudResource, []voc.IsAnomalyDetection) {
-	var (
-		list                 []voc.IsCloudResource
-		anomalyDetectionList []voc.IsAnomalyDetection
-		err                  error
-	)
-
-	// initialize SQL databases client
-	if err = d.initDatabasesClient(); err != nil {
-		log.Errorf("error initializing database client: %v", err)
-		return list, anomalyDetectionList
-	}
-
-	// Get databases for given server
-	serverlistPager := d.clients.databasesClient.NewListByServerPager(resourceGroupName(util.Deref(server.ID)), *server.Name, &armsql.DatabasesClientListByServerOptions{})
-	for serverlistPager.More() {
-		pageResponse, err := serverlistPager.NextPage(context.TODO())
-		if err != nil {
-			log.Errorf("%s: %v", ErrGettingNextPage, err)
-			return list, anomalyDetectionList
-		}
-
-		for _, value := range pageResponse.Value {
-			// Create anomaly detection property
-			// Get anomaly detection status
-			anomalyDetectionEnabled, err := d.anomalyDetectionEnabled(server, value)
-			if err != nil {
-				log.Errorf("error getting anomaly detection info for database '%s': %v", *value.Name, err)
-			}
-
-			a := &voc.AnomalyDetection{
-				Scope:   voc.ResourceID(*value.ID),
-				Enabled: anomalyDetectionEnabled,
-			}
-
-			anomalyDetectionList = append(anomalyDetectionList, a)
-
-			// Create database storage voc object
-			sqlDB := &voc.DatabaseStorage{
-				Storage: &voc.Storage{
-					Resource: discovery.NewResource(d,
-						voc.ResourceID(*value.ID),
-						*value.Name,
-						value.Properties.CreationDate,
-						voc.GeoLocation{
-							Region: *value.Location,
-						},
-						labels(value.Tags),
-						voc.ResourceID(*server.ID),
-						voc.DatabaseStorageType,
-						value),
-					AtRestEncryption: &voc.AtRestEncryption{
-						Enabled:   *value.Properties.IsInfraEncryptionEnabled,
-						Algorithm: AES256,
-					},
-					// TODO(all): Backups
-				},
-			}
-			list = append(list, sqlDB)
-		}
-	}
-
-	return list, anomalyDetectionList
-}
-
-// checkTlsVersion returns Clouditor's TLS version constants for the given TLS version
-func checkTlsVersion(version string) string {
-	// Check TLS version
-	switch version {
-	case constants.TLS1_0, constants.TLS1_1, constants.TLS1_2:
-		return version
-	case "1.0", "1_0":
-		return constants.TLS1_0
-	case "1.1", "1_1":
-		return constants.TLS1_1
-	case "1.2", "1_2":
-		return constants.TLS1_2
-	default:
-		log.Warningf("'%s' is no implemented TLS version.", version)
-		return ""
-	}
-}
-
-func (d *azureDiscovery) discoverStorageAccounts() ([]voc.IsCloudResource, error) {
-	var storageResourcesList []voc.IsCloudResource
-
-	// initialize backup policies client
-	if err := d.initBackupPoliciesClient(); err != nil {
-		return nil, err
-	}
-
-	// initialize backup instances client
-	if err := d.initBackupInstancesClient(); err != nil {
-		return nil, err
-	}
-
-	// initialize backup vaults client
-	if err := d.initBackupVaultsClient(); err != nil {
-		return nil, err
-	}
-
-	// initialize storage accounts client
-	if err := d.initAccountsClient(); err != nil {
-		return nil, err
-	}
-
-	// initialize blob container client
-	if err := d.initBlobContainerClient(); err != nil {
-		return nil, err
-	}
-
-	// initialize file share client
-	if err := d.initFileStorageClient(); err != nil {
-		return nil, err
-	}
-
-	// Discover backup vaults
-	err := d.discoverBackupVaults()
-	if err != nil {
-		log.Errorf("could not discover backup vaults: %v", err)
-	}
-
-	// Discover object and file storages
-	err = listPager(d,
-		d.clients.accountsClient.NewListPager,
-		d.clients.accountsClient.NewListByResourceGroupPager,
-		func(res armstorage.AccountsClientListResponse) []*armstorage.Account {
-			return res.Value
-		},
-		func(res armstorage.AccountsClientListByResourceGroupResponse) []*armstorage.Account {
-			return res.Value
-		},
-		func(account *armstorage.Account) error {
-			// Discover object storages
-			objectStorages, err := d.discoverObjectStorages(account)
-			if err != nil {
-				return fmt.Errorf("could not handle object storages: %w", err)
-			}
-
-			// Discover file storages
-			fileStorages, err := d.discoverFileStorages(account)
-			if err != nil {
-				return fmt.Errorf("could not handle file storages: %w", err)
-			}
-
-			storageResourcesList = append(storageResourcesList, objectStorages...)
-			storageResourcesList = append(storageResourcesList, fileStorages...)
-
-			// Create storage service for all storage account resources
-			storageService, err := d.handleStorageAccount(account, storageResourcesList)
-			if err != nil {
-				return fmt.Errorf("could not create storage service: %w", err)
-			}
-
-			storageResourcesList = append(storageResourcesList, storageService)
-
-			return nil
-		})
-	if err != nil {
-		return nil, err
-	}
-
-	// Add backup storage account objects
-	if d.backupMap[DataSourceTypeStorageAccountObject] != nil && d.backupMap[DataSourceTypeStorageAccountObject].backupStorages != nil {
-		storageResourcesList = append(storageResourcesList, d.backupMap[DataSourceTypeStorageAccountObject].backupStorages...)
-	}
-
-	return storageResourcesList, nil
-}
-
-func (d *azureDiscovery) discoverFileStorages(account *armstorage.Account) ([]voc.IsCloudResource, error) {
-	var list []voc.IsCloudResource
-
-	// List all file shares in the specified resource group
-	listPager := d.clients.fileStorageClient.NewListPager(resourceGroupName(util.Deref(account.ID)), util.Deref(account.Name), &armstorage.FileSharesClientListOptions{})
-	for listPager.More() {
-		pageResponse, err := listPager.NextPage(context.TODO())
-		if err != nil {
-			err = fmt.Errorf("%s: %v", ErrGettingNextPage, err)
-			return nil, err
-		}
-
-		for _, value := range pageResponse.Value {
-			fileStorages, err := d.handleFileStorage(account, value)
-			if err != nil {
-				return nil, fmt.Errorf("could not handle file storage: %w", err)
-			}
-
-			log.Infof("Adding file storage '%s", fileStorages.Name)
-
-			list = append(list, fileStorages)
-		}
-	}
-
-	return list, nil
-}
-
-func (d *azureDiscovery) discoverObjectStorages(account *armstorage.Account) ([]voc.IsCloudResource, error) {
-	var list []voc.IsCloudResource
-
-	// List all blob containers in the specified resource group
-	listPager := d.clients.blobContainerClient.NewListPager(resourceGroupName(util.Deref(account.ID)), util.Deref(account.Name), &armstorage.BlobContainersClientListOptions{})
-	for listPager.More() {
-		pageResponse, err := listPager.NextPage(context.TODO())
-		if err != nil {
-			err = fmt.Errorf("%s: %v", ErrGettingNextPage, err)
-			return nil, err
-		}
-
-		for _, value := range pageResponse.Value {
-			objectStorages, err := d.handleObjectStorage(account, value)
-			if err != nil {
-				return nil, fmt.Errorf("could not handle object storage: %w", err)
-			}
-			log.Infof("Adding object storage '%s'", objectStorages.Name)
-
-			list = append(list, objectStorages)
-
-		}
-	}
-
-	return list, nil
-}
-
-func (d *azureDiscovery) handleStorageAccount(account *armstorage.Account, storagesList []voc.IsCloudResource) (*voc.ObjectStorageService, error) {
-	var (
-		storageResourceIDs []voc.ResourceID
-	)
-
-	if account == nil {
-		return nil, ErrEmptyStorageAccount
-	}
-
-	// Get all object storage IDs
-	for _, storage := range storagesList {
-		if strings.Contains(string(storage.GetID()), accountName(util.Deref(account.ID))) {
-			storageResourceIDs = append(storageResourceIDs, storage.GetID())
-		}
-	}
-
-	te := &voc.TransportEncryption{
-		Enforced:   util.Deref(account.Properties.EnableHTTPSTrafficOnly),
-		Enabled:    true, // cannot be disabled
-		TlsVersion: checkTlsVersion((string(util.Deref(account.Properties.MinimumTLSVersion)))),
-		Algorithm:  constants.TLS,
-	}
-
-	storageService := &voc.ObjectStorageService{
-		StorageService: &voc.StorageService{
-			Storage: storageResourceIDs,
-			NetworkService: &voc.NetworkService{
-				Networking: &voc.Networking{
-					Resource: discovery.NewResource(d,
-						voc.ResourceID(util.Deref(account.ID)),
-						util.Deref(account.Name),
-						account.Properties.CreationTime,
-						voc.GeoLocation{
-							Region: util.Deref(account.Location),
-						},
-						labels(account.Tags),
-						resourceGroupID(account.ID),
-						voc.ObjectStorageServiceType,
-						account,
-					),
-				},
-				TransportEncryption: te,
-			},
-		},
-		HttpEndpoint: &voc.HttpEndpoint{
-			Url:                 generalizeURL(util.Deref(account.Properties.PrimaryEndpoints.Blob)),
-			TransportEncryption: te,
-		},
-	}
-
-	return storageService, nil
-}
-
-func (d *azureDiscovery) handleFileStorage(account *armstorage.Account, fileshare *armstorage.FileShareItem) (*voc.FileStorage, error) {
-	var (
-		monitoringLogDataEnabled bool
-		securityAlertsEnabled    bool
-	)
-
-	if account == nil {
-		return nil, ErrEmptyStorageAccount
-	}
-
-	// It is possible that the fileshare is not empty. In that case we have to check if a mandatory field is empty, so the whole disk is empty
-	if fileshare == nil || fileshare.ID == nil {
-		return nil, fmt.Errorf("fileshare is nil")
-	}
-
-	// Get atRestEncryptionEnabled
-	enc, err := storageAtRestEncryption(account)
-	if err != nil {
-		return nil, fmt.Errorf("could not get file storage properties for the atRestEncryption: %w", err)
-	}
-
-	// Get monitoringLogDataEnabled and securityAlertsEnabled
-	if d.defenderProperties[DefenderStorageType] != nil {
-		monitoringLogDataEnabled = d.defenderProperties[DefenderVirtualMachineType].monitoringLogDataEnabled
-		securityAlertsEnabled = d.defenderProperties[DefenderVirtualMachineType].securityAlertsEnabled
-	}
-
-	return &voc.FileStorage{
-		Storage: &voc.Storage{
-			Resource: discovery.NewResource(d,
-				voc.ResourceID(util.Deref(fileshare.ID)),
-				util.Deref(fileshare.Name),
-				// We only have the creation time of the storage account the file storage belongs to
-				account.Properties.CreationTime,
-				voc.GeoLocation{
-					// The location is the same as the storage account
-					Region: util.Deref(account.Location),
-				},
-				// The storage account labels the file storage belongs to
-				labels(account.Tags),
-				// the storage account is our parent
-				voc.ResourceID(util.Deref(account.ID)),
-				voc.FileStorageType,
-				account, fileshare,
-			),
-			ResourceLogging: &voc.ResourceLogging{
-				Logging: &voc.Logging{
-					MonitoringLogDataEnabled: monitoringLogDataEnabled,
-					SecurityAlertsEnabled:    securityAlertsEnabled,
-				},
-			},
-			AtRestEncryption: enc,
-		},
-	}, nil
-}
-
-func (d *azureDiscovery) handleObjectStorage(account *armstorage.Account, container *armstorage.ListContainerItem) (*voc.ObjectStorage, error) {
-	var (
-		backups                  []*voc.Backup
-		monitoringLogDataEnabled bool
-		securityAlertsEnabled    bool
-	)
-
-	if account == nil {
-		return nil, ErrEmptyStorageAccount
-	}
-
-	// It is possible that the container is not empty. In that case we have to check if a mandatory field is empty, so the whole disk is empty
-	if container == nil || container.ID == nil {
-		return nil, fmt.Errorf("container is nil")
-	}
-
-	enc, err := storageAtRestEncryption(account)
-	if err != nil {
-		return nil, fmt.Errorf("could not get object storage properties for the atRestEncryption: %w", err)
-	}
-
-	if d.backupMap[DataSourceTypeStorageAccountObject] != nil && d.backupMap[DataSourceTypeStorageAccountObject].backup[util.Deref(account.ID)] != nil {
-		backups = d.backupMap[DataSourceTypeStorageAccountObject].backup[util.Deref(account.ID)]
-	}
-	backups = backupsEmptyCheck(backups)
-
-	if d.defenderProperties[DefenderStorageType] != nil {
-		monitoringLogDataEnabled = d.defenderProperties[DefenderVirtualMachineType].monitoringLogDataEnabled
-		securityAlertsEnabled = d.defenderProperties[DefenderVirtualMachineType].securityAlertsEnabled
-	}
-
-	return &voc.ObjectStorage{
-		Storage: &voc.Storage{
-			Resource: discovery.NewResource(d,
-				voc.ResourceID(util.Deref(container.ID)),
-				util.Deref(container.Name),
-				// We only have the creation time of the storage account the object storage belongs to
-				account.Properties.CreationTime,
-				voc.GeoLocation{
-					// The location is the same as the storage account
-					Region: util.Deref(account.Location),
-				},
-				// The storage account labels the object storage belongs to
-				labels(account.Tags),
-				// the storage account is our parent
-				voc.ResourceID(util.Deref(account.ID)),
-				voc.ObjectStorageType,
-				account, container,
-			),
-			AtRestEncryption: enc,
-			Immutability: &voc.Immutability{
-				Enabled: util.Deref(container.Properties.HasImmutabilityPolicy),
-			},
-			ResourceLogging: &voc.ResourceLogging{
-				Logging: &voc.Logging{
-					MonitoringLogDataEnabled: monitoringLogDataEnabled,
-					SecurityAlertsEnabled:    securityAlertsEnabled,
-				},
-			},
-			Backups: backups,
-		},
-		PublicAccess: util.Deref(container.Properties.PublicAccess) != armstorage.PublicAccessNone,
-	}, nil
-}
-
-// storageAtRestEncryption takes encryption properties of an armstorage.Account and converts it into our respective ontology object.
-func storageAtRestEncryption(account *armstorage.Account) (enc voc.IsAtRestEncryption, err error) {
-	if account == nil {
-		return enc, ErrEmptyStorageAccount
-	}
-
-	if account.Properties == nil || account.Properties.Encryption.KeySource == nil {
-		return enc, errors.New("keySource is empty")
-	} else if util.Deref(account.Properties.Encryption.KeySource) == armstorage.KeySourceMicrosoftStorage {
-		enc = &voc.ManagedKeyEncryption{
-			AtRestEncryption: &voc.AtRestEncryption{
-				Algorithm: AES256,
-				Enabled:   true,
-			},
-		}
-	} else if util.Deref(account.Properties.Encryption.KeySource) == armstorage.KeySourceMicrosoftKeyvault {
-		enc = &voc.CustomerKeyEncryption{
-			AtRestEncryption: &voc.AtRestEncryption{
-				Algorithm: "", // TODO(all): TBD
-				Enabled:   true,
-			},
-			KeyUrl: util.Deref(account.Properties.Encryption.KeyVaultProperties.KeyVaultURI),
-		}
-	}
-
-	return enc, nil
-}
-
-// anomalyDetectionEnabled returns true if Azure Advanced Threat Protection is enabled for the database.
-func (d *azureDiscovery) anomalyDetectionEnabled(server *armsql.Server, db *armsql.Database) (bool, error) {
-	// initialize threat protection client
-	if err := d.initThreatProtectionClient(); err != nil {
-		return false, err
-	}
-
-	listPager := d.clients.threatProtectionClient.NewListByDatabasePager(resourceGroupName(util.Deref(db.ID)), *server.Name, *db.Name, &armsql.DatabaseAdvancedThreatProtectionSettingsClientListByDatabaseOptions{})
-	for listPager.More() {
-		pageResponse, err := listPager.NextPage(context.TODO())
-		if err != nil {
-			err = fmt.Errorf("%s: %v", ErrGettingNextPage, err)
-			return false, err
-		}
-
-		for _, value := range pageResponse.Value {
-			if *value.Properties.State == armsql.AdvancedThreatProtectionStateEnabled {
-				return true, nil
-			}
-		}
-	}
-	return false, nil
-}
-
-// accountName return the ID's account name
-func accountName(id string) string {
-	if id == "" {
-		return ""
-	}
-
-	splitName := strings.Split(id, "/")
-	return splitName[8]
-}
-
-// generalizeURL generalizes the URL, because the URL depends on the storage type
-func generalizeURL(url string) string {
-	if url == "" {
-		return ""
-	}
-
-	urlSplit := strings.Split(url, ".")
-	urlSplit[1] = "[file,blob]"
-	newURL := strings.Join(urlSplit, ".")
-
-	return newURL
-}
